requirements.txt
torch==2.4.0
torchvision==0.19.0
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
transformers==4.35.2
huggingface-hub==0.19.4
matplotlib==3.7.2 